MC-31983: Refactor code to match the agreed repo module structure
- Fixed CustomerStorefrontGraphQl namepsaces
